+++ runtime/src/main/java/com/fizzed/rocker/runtime/DefaultRockerModel.java
@@ -18,12 +18,12 @@ package com.fizzed.rocker.runtime;
 import com.fizzed.rocker.RenderingException;
 import com.fizzed.rocker.RockerOutput;
 import com.fizzed.rocker.RockerModel;
-import com.fizzed.rocker.RockerModelCallback;
+import com.fizzed.rocker.RockerOutputFactory;
+import com.fizzed.rocker.RockerTemplateCustomizer;
 
 public class DefaultRockerModel<T extends DefaultRockerModel> implements RockerModel {
     
     private boolean rendered;
-    private RockerModelCallback callback;
     
     /**
      * Renders model and template content to output. Single use only.
@@ -34,10 +34,23 @@ public class DefaultRockerModel<T extends DefaultRockerModel> implements RockerM
      */
     @Override
     public RockerOutput render() throws RenderingException {
-        return render(null);
+        return doRender(null, null, null);
     }
     
-    protected RockerOutput render(DefaultRockerTemplate context) throws RenderingException {
+    @Override
+    public <O extends RockerOutput> O render(RockerOutputFactory<O> outputFactory) throws RenderingException {
+        return (O)doRender(null, outputFactory, null);
+    }
+    
+    @Override
+    public <O extends RockerOutput> O render(RockerOutputFactory<O> outputFactory,
+                                             RockerTemplateCustomizer templateCustomizer) throws RenderingException {
+        return (O)doRender(null, outputFactory, templateCustomizer);
+    }
+    
+    protected RockerOutput doRender(DefaultRockerTemplate context,
+                                    RockerOutputFactory outputFactory,
+                                    RockerTemplateCustomizer templateCustomizer) throws RenderingException {
         // no real need for thread safety since templates should only be used by a single thread
         if (this.rendered) {
             throw new RenderingException("Template already rendered (templates are single use only!)");
@@ -45,21 +58,17 @@ public class DefaultRockerModel<T extends DefaultRockerModel> implements RockerM
         
         DefaultRockerTemplate template = buildTemplate();
         
-        if (this.callback != null) {
-            this.callback.onRender(template);
+        if (templateCustomizer != null) {
+            templateCustomizer.customize(template);
         }
         
         this.rendered = true;
         
-        return template.__render(context);
+        return template.__render(context, outputFactory);
     }
     
     protected DefaultRockerTemplate buildTemplate() throws RenderingException  {
         throw new RenderingException("Rocker model does not implement buildTemplate(). Did you forget to implement?");
-    }
-    
-    public void __callback(RockerModelCallback callback) {
-        this.callback = callback;
     }
     
     @Override