+++ runtime/src/test/java/com/fizzed/rocker/runtime/DefaultRockerTemplateTest.java
@@ -18,8 +18,10 @@ package com.fizzed.rocker.runtime;
 import com.fizzed.rocker.ContentType;
 import com.fizzed.rocker.RenderingException;
 import com.fizzed.rocker.RockerOutput;
+import com.fizzed.rocker.RockerOutputFactory;
 import java.io.IOException;
 import static org.hamcrest.CoreMatchers.containsString;
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
@@ -106,6 +108,41 @@ public class DefaultRockerTemplateTest {
         RockerOutput out = model.render();
         
         assertThat(out.toString(), is("Hello!"));
+        assertThat(out, instanceOf(ArrayOfByteArraysOutput.class));
+    }
+    
+    @Test
+    public void customOutput() throws Exception {
+        // bare minimum model + template
+        DefaultRockerModel model = new DefaultRockerModel() {
+            @Override
+            protected DefaultRockerTemplate buildTemplate() throws RenderingException {
+                return new DefaultRockerTemplate(this) {
+                    // anonymous initializer!
+                    {
+                        this.__internal.setCharset("UTF-8");
+                        this.__internal.setContentType(ContentType.HTML);
+                    }
+                    
+                    @Override
+                    protected void __doRender() throws IOException, RenderingException {
+                        this.__internal.writeValue("Hello!");
+                    }
+                };
+            }
+        };
+        
+        RockerOutputFactory<StringBuilderOutput> outputFactory = new RockerOutputFactory<StringBuilderOutput>() {
+            @Override
+            public StringBuilderOutput create(ContentType contentType, String charsetName) {
+                return new StringBuilderOutput(contentType, charsetName);
+            }
+        };
+        
+        RockerOutput out = model.render(outputFactory);
+        
+        assertThat(out.toString(), is("Hello!"));
+        assertThat(out, instanceOf(StringBuilderOutput.class));
     }